+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * add a filter to Plugin icons table.
 * rework `.htaccess` files (PNG icon to SVG).
 * rework `kubernetes*.yml` files (PNG icon to SVG).
+* rework `*.toml` files (PNG icon to SVG).
 
 ### 1.32.0 (2020/07/11)
 * improve VERSION files detection.